+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/peertask/task/GetSyncReceiptsFromPeerTask.java
@@ -36,6 +36,7 @@ import org.hyperledger.besu.ethereum.p2p.rlpx.wire.MessageData;
 import org.hyperledger.besu.ethereum.p2p.rlpx.wire.SubProtocol;
 import org.hyperledger.besu.ethereum.rlp.RLPException;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -45,32 +46,30 @@ import java.util.function.Predicate;
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.tuweni.bytes.Bytes;
 
-public class GetSyncReceiptsFromPeerTask
-    implements PeerTask<Map<SyncBlock, List<SyncTransactionReceipt>>> {
-
-  private final List<SyncBlock> requestedBlocks;
+public class GetSyncReceiptsFromPeerTask implements PeerTask<GetSyncReceiptsFromPeerTask.Response> {
+  private final Request request;
+  protected final ProtocolSchedule protocolSchedule;
   private final List<BlockHeader> requestedHeaders;
   private final long requiredBlockchainHeight;
   private final boolean isPoS;
   private final SyncTransactionReceiptEncoder syncTransactionReceiptEncoder;
 
   public GetSyncReceiptsFromPeerTask(
-      final List<SyncBlock> blocks,
+      final Request request,
       final ProtocolSchedule protocolSchedule,
       final SyncTransactionReceiptEncoder syncTransactionReceiptEncoder) {
-    checkArgument(!blocks.isEmpty(), "Requested block list must not be empty");
-    this.requestedBlocks = blocks;
-    this.requestedHeaders = blocks.stream().map(SyncBlock::getHeader).toList();
+    checkArgument(!request.blocks.isEmpty(), "Requested block list must not be empty");
+    this.request = request;
+    this.protocolSchedule = protocolSchedule;
+    this.requestedHeaders = request.blocks.stream().map(SyncBlock::getHeader).toList();
 
-    // calculate the minimum required blockchain height a peer will need to be able to fulfil this
-    // request
     requiredBlockchainHeight =
-        requestedHeaders.stream()
+        this.requestedHeaders.stream()
             .mapToLong(BlockHeader::getNumber)
             .max()
             .orElse(BlockHeader.GENESIS_BLOCK_NUMBER);
 
-    isPoS = protocolSchedule.anyMatch((ps) -> ps.spec().isPoS());
+    isPoS = protocolSchedule.anyMatch(ps -> ps.spec().isPoS());
     this.syncTransactionReceiptEncoder = syncTransactionReceiptEncoder;
   }
 
@@ -81,11 +80,14 @@ public class GetSyncReceiptsFromPeerTask
 
   @Override
   public MessageData getRequestMessage(final Set<Capability> agreedCapabilities) {
-    return GetReceiptsMessage.create(requestedHeaders.stream().map(BlockHeader::getHash).toList());
+    final List<Hash> blockHashes = requestedHeaders.stream().map(BlockHeader::getHash).toList();
+    return agreedCapabilities.stream().anyMatch(EthProtocol::isEth70Compatible)
+        ? GetReceiptsMessage.create(blockHashes, request.firstBlockPartialReceipts.size())
+        : GetReceiptsMessage.create(blockHashes);
   }
 
   @Override
-  public Map<SyncBlock, List<SyncTransactionReceipt>> processResponse(
+  public Response processResponse(
       final MessageData messageData, final Set<Capability> agreedCapabilities)
       throws InvalidPeerTaskResponseException, MalformedRlpFromPeerException {
     if (messageData == null) {
@@ -93,20 +95,70 @@ public class GetSyncReceiptsFromPeerTask
     }
     final ReceiptsMessage receiptsMessage = ReceiptsMessage.readFrom(messageData);
     try {
-      final List<List<SyncTransactionReceipt>> receivedBlocks = receiptsMessage.syncReceipts();
-      if (receivedBlocks.size() > requestedBlocks.size()) {
+      final boolean isEth70Response = receiptsMessage.lastBlockIncomplete().isPresent();
+
+      final List<List<SyncTransactionReceipt>> receivedReceipts =
+          isEth70Response
+              ? completeFirstBlock(receiptsMessage.syncReceipts())
+              : receiptsMessage.syncReceipts();
+
+      if (receivedReceipts.isEmpty()) {
+        throw new InvalidPeerTaskResponseException("No result returned");
+      }
+
+      if (receivedReceipts.size() > request.size()) {
         throw new InvalidPeerTaskResponseException("Too many result returned");
       }
-      final Map<SyncBlock, List<SyncTransactionReceipt>> response =
-          HashMap.newHashMap(receivedBlocks.size());
-      for (int i = 0; i < receivedBlocks.size(); i++) {
-        response.put(requestedBlocks.get(i), receivedBlocks.get(i));
+
+      final int endIndex;
+      final List<SyncTransactionReceipt> lastBlockPartialReceipts;
+      if (isEth70Response && receiptsMessage.lastBlockIncomplete().get()) {
+        endIndex = receivedReceipts.size() - 1;
+        lastBlockPartialReceipts = receivedReceipts.getLast();
+      } else {
+        endIndex = receivedReceipts.size();
+        lastBlockPartialReceipts = List.of();
       }
-      return response;
+
+      final Map<SyncBlock, List<SyncTransactionReceipt>> receiptsByBlock =
+          HashMap.newHashMap(receivedReceipts.size());
+
+      for (int i = 0; i < endIndex; i++) {
+        receiptsByBlock.put(request.blocks.get(i), receivedReceipts.get(i));
+      }
+
+      return new Response(receiptsByBlock, lastBlockPartialReceipts);
     } catch (RLPException e) {
       // indicates a malformed or unexpected RLP result from the peer
       throw new MalformedRlpFromPeerException(e, messageData.getData());
     }
+  }
+
+  private List<List<SyncTransactionReceipt>> completeFirstBlock(
+      final List<List<SyncTransactionReceipt>> receivedReceipts)
+      throws InvalidPeerTaskResponseException {
+    if (request.firstBlockPartialReceipts.isEmpty()) {
+      // nothing to integrate returning as is
+      return receivedReceipts;
+    }
+
+    if (receivedReceipts.isEmpty()) {
+      throw new InvalidPeerTaskResponseException("No result returned");
+    }
+
+    // add new receipts to the already present ones
+    final List<SyncTransactionReceipt> cumulativeReceiptsForFirstBlock =
+        new ArrayList<>(
+            request.firstBlockPartialReceipts.size() + receivedReceipts.getFirst().size());
+    cumulativeReceiptsForFirstBlock.addAll(request.firstBlockPartialReceipts);
+    cumulativeReceiptsForFirstBlock.addAll(receivedReceipts.getFirst());
+
+    // replace first list of receipts with the new cumulative list
+    final List<List<SyncTransactionReceipt>> cumulativeReceipts =
+        new ArrayList<>(receivedReceipts.size());
+    cumulativeReceipts.add(cumulativeReceiptsForFirstBlock);
+    cumulativeReceipts.addAll(receivedReceipts.subList(1, receivedReceipts.size()));
+    return cumulativeReceipts;
   }
 
   @Override
@@ -115,13 +167,13 @@ public class GetSyncReceiptsFromPeerTask
   }
 
   @Override
-  public PeerTaskValidationResponse validateResult(
-      final Map<SyncBlock, List<SyncTransactionReceipt>> result) {
+  public PeerTaskValidationResponse validateResult(final Response result) {
     if (result.isEmpty()) {
       return PeerTaskValidationResponse.NO_RESULTS_RETURNED;
     }
 
-    for (final Map.Entry<SyncBlock, List<SyncTransactionReceipt>> entry : result.entrySet()) {
+    for (final Map.Entry<SyncBlock, List<SyncTransactionReceipt>> entry :
+        result.completeReceiptsByBlock.entrySet()) {
       final SyncBlock requestedBlock = entry.getKey();
       final List<SyncTransactionReceipt> receivedReceiptsForBlock = entry.getValue();
 
@@ -136,12 +188,21 @@ public class GetSyncReceiptsFromPeerTask
       }
     }
 
+    if (!result.lastBlockPartialReceipts().isEmpty()) {
+      final SyncBlock lastBlockReceived =
+          request.blocks.get(result.completeReceiptsByBlock().size());
+      if (result.lastBlockPartialReceipts().size()
+          > lastBlockReceived.getBody().getTransactionCount()) {
+        return PeerTaskValidationResponse.TOO_MANY_RESULTS_RETURNED;
+      }
+    }
+
     return PeerTaskValidationResponse.RESULTS_VALID_AND_GOOD;
   }
 
   private boolean receiptsRootMatches(
       final BlockHeader blockHeader, final List<SyncTransactionReceipt> receipts) {
-    final Hash calculatedReceiptsRoot =
+    final var calculatedReceiptsRoot =
         Util.getRootFromListOfBytes(
             receipts.stream()
                 .map(
@@ -160,6 +221,29 @@ public class GetSyncReceiptsFromPeerTask
 
   @VisibleForTesting
   public List<SyncBlock> getRequestedBlocks() {
-    return requestedBlocks;
+    return request.blocks();
+  }
+
+  public record Request(
+      List<SyncBlock> blocks, List<SyncTransactionReceipt> firstBlockPartialReceipts) {
+    public boolean isEmpty() {
+      return blocks.isEmpty();
+    }
+
+    public int size() {
+      return blocks.size();
+    }
+  }
+
+  public record Response(
+      Map<SyncBlock, List<SyncTransactionReceipt>> completeReceiptsByBlock,
+      List<SyncTransactionReceipt> lastBlockPartialReceipts) {
+    public boolean isEmpty() {
+      return completeReceiptsByBlock.isEmpty() && lastBlockPartialReceipts.isEmpty();
+    }
+
+    public int size() {
+      return completeReceiptsByBlock.size();
+    }
   }
 }